pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentforge"
version = "1.0.0"
description = "Declarative AI Agent Skill Framework — Build agent skills with YAML, not code"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "AgentForge Contributors"}]
keywords = ["ai", "agent", "skill", "framework", "llm", "automation", "yaml"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.28",
    "rich>=13.0",
    "click>=8.0",
    "jinja2>=3.0",
]

[project.optional-dependencies]
full = ["openai>=1.0", "anthropic>=0.20", "feedparser>=6.0"]
dev = ["pytest>=7.0", "black", "ruff"]

[project.scripts]
agentforge = "agentforge.cli:main"
af = "agentforge.cli:main"

[tool.setuptools.packages.find]
include = ["agentforge*"]

[tool.setuptools.package-data]
agentforge = ["skills/*.yaml", "skills/*.yml"]

[project.urls]
Homepage = "https://github.com/dnsrootserver/www"
Documentation = "https://github.com/dnsrootserver/www#readme"
Repository = "https://github.com/dnsrootserver/www"
